LayoutTests/fast/filesystem/file-entry-to-uri-expected.txt

Obtaining URI from FileEntry.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testFileURI is 'filesystem:file:///temporary/testFileEntry.txt'
PASS successfullyParsed is true

TEST COMPLETE
